requirements.txt
langchain==0.2.11
langchain_community==0.2.10
numpy==1.24.4
openai==1.37.1
pandas==2.0.3
pdf2image==1.17.0
Pillow==10.4.0
streamlit==1.37.0
